Internals

Others

EvoLinear.Linear.update_∇! — Function

update_∇!(L, ∇¹, ∇², x, y, p, w)

Update gradients w.r.t each feature. Each feature gradient update is dispatch according to the loss type (mse, logistic...).

EvoLinear.Linear.predict_linear — Function

predict_linear(m, x)

Returns the predictions on the linear basis from model m using the features matrix x.

Arguments

m::EvoLinearModel: model generating the predictions.

x: features matrix [nobs, num_features] for which predictions are generated.

EvoLinear.Linear.predict_proj — Function

predict_proj(m, x)

Returns the predictions on the projected basis from model m using the features matrix x.

MSE: pred_proj = pred_linear

Logistic: pred_proj = sigmoid(pred_linear)

Poisson: pred_proj = exp(pred_linear)

Gamma: pred_proj = exp(pred_linear)

Tweedie: pred_proj = exp(pred_linear)

Arguments

m::EvoLinearModel: model generating the predictions.

x: features matrix [nobs, num_features] for which predictions are generated.