API

Regressors

EvoLinear.EvoLinearRegressor — Type

EvoLinearRegressor(; kwargs...)

loss=:mse: loss function to be minimised. Can be one of:

:mse

:logistic

:poisson

:gamma

:tweedie

nrounds=10: maximum number of training rounds.

eta=1: Learning rate. Typically in the range [1e-2, 1].

L1=0: Regularization penalty applied by shrinking to 0 weight update if update is < L1. No penalty if update > L1. Results in sparse feature selection. Typically in the [0, 1] range on normalized features.

L2=0: Regularization penalty applied to the squared of the weight update value. Restricts large parameter values. Typically in the [0, 1] range on normalized features.

rng=123: random seed. Not used at the moment.

updater=:all: training method. Only :all is supported at the moment. Gradients for each feature are computed simultaneously, then bias is updated based on all features update.

device=:cpu: Only :cpu is supported at the moment.

Training & Inference

EvoLinear.fit — Function

fit(config::EvoLinearRegressor;
    x, y, w=nothing,
    x_eval=nothing, y_eval=nothing, w_eval=nothing,
    metric=:none,
    print_every_n=1,
    tol=1e-5)

Provided a config, EvoLinear.fit takes x and y as features and target inputs, plus optionally w as weights and train a Linear boosted model.

Arguments

config::EvoLinearRegressor:

Keyword arguments

x::AbstractMatrix: Features matrix. Dimensions are [nobs, num_features].

y::AbstractVector: Vector of observed targets.

w=nothing: Vector of weights. Can be be either a Vector or nothing. If nothing, assumes a vector of 1s.

metric=:none: Evaluation metric to be tracked through each iteration. Can be one of:

:mse

:logistic

:poisson_deviance

:gamma_deviance

:tweedie_deviance

EvoLinear.predict_linear — Function

predict_linear(m, x)

Returns the predictions on the linear basis from model m using the features matrix x.

Arguments

m::EvoLinearModel

x

EvoLinear.predict_proj — Function

predict_proj(m, x)

Returns the predictions on the projected basis from model m using the features matrix x.

MSE: predproj = predlinear

Logistic: predproj = sigmoid(predlinear)

Poisson: predproj = exp(predlinear)

Gamma: predproj = exp(predlinear)

Tweedie: predproj = exp(predlinear)

Arguments

m::EvoLinearModel

x

Evaluation metrics

EvoLinear.mse — Function

mse(p, y) mse(p, y, w)

Mean squared error evaluation metric.

Arguments

p: predicted value.

y: observed target variable.

w: vector of weights.

EvoLinear.mae — Function

mae(p, y)
mae(p, y, w)

Mean absolute error evaluation metric.

Arguments

p: predicted value.

y: observed target variable.

w: vector of weights.

EvoLinear.logloss — Function

logloss(p, y)
logloss(p, y, w)

Logloss evaluation metric. ylog(p) + (1-y)log(1-p)

Arguments

p: predicted value. Assumes that p is on a projected basis (ie. in the [0-1] range).

y: observed target variable.

w: vector of weights.

EvoLinear.poisson_deviance — Function

poisson_deviance(p, y)
poisson_deviance(p, y, w)

Poisson deviance evaluation metric. 𝐷 = 2 * (y * log(y/p) + p - y)

Arguments

p: predicted value. Assumes that p is on a projected basis (ie. in the [0-Inf] range).

y: observed target variable.

w: vector of weights.

EvoLinear.gamma_deviance — Function

gamma_deviance(p, y)
gamma_deviance(p, y, w)

Gamma deviance evaluation metric. 𝐷 = 2 * (log(μ/y) + y/μ - 1)

Arguments

p: predicted value. Assumes that p is on a projected basis (ie. in the [0-Inf] range).

y: observed target variable.

w: vector of weights.

EvoLinear.tweedie_deviance — Function

tweedie_deviance(p, y)
tweedie_deviance(p, y, w)

Tweedie deviance evaluation metric. Fixed rho (ρ) of 1.5. 𝐷 = 2 * (y²⁻ʳʰᵒ/(1-rho)(2-rho) - yμ¹⁻ʳʰᵒ/(1-rho) + μ²⁻ʳʰᵒ/(2-rho))

Arguments

p: predicted value. Assumes that p is on a projected basis (ie. in the [0-Inf] range).

y: observed target variable.

w: vector of weights.